backend/baking_app/requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
sqlparse==0.4.4